+++ Demo/src/main/java/com/paypal/android/ui/vaultcard/VaultCardView.kt
@@ -20,19 +20,15 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.paypal.android.R
-import com.paypal.android.cardpayments.CardAuthChallenge
 import com.paypal.android.cardpayments.CardVaultResult
 import com.paypal.android.uishared.components.ActionButtonColumn
 import com.paypal.android.uishared.components.CardForm
 import com.paypal.android.uishared.components.CardPaymentTokenView
 import com.paypal.android.uishared.components.CardSetupTokenView
-import com.paypal.android.uishared.components.CardVaultResultView
 import com.paypal.android.uishared.components.EnumOptionList
 import com.paypal.android.uishared.components.ErrorView
-import com.paypal.android.uishared.components.InfoColumn
 import com.paypal.android.uishared.components.PropertyView
 import com.paypal.android.uishared.components.StepHeader
-import com.paypal.android.uishared.state.ActionState
 import com.paypal.android.uishared.state.CompletedActionState
 import com.paypal.android.utils.OnLifecycleOwnerResumeEffect
 import com.paypal.android.utils.UIConstants
@@ -66,27 +62,13 @@ fun VaultCardView(
             .padding(horizontal = contentPadding)
             .verticalScroll(scrollState)
     ) {
-        Step_CreateSetupToken(stepNumber = 1, uiState, viewModel)
+        Step_CreateSetupToken(uiState, viewModel)
         if (uiState.isCreateSetupTokenSuccessful) {
-            Step_VaultCard(stepNumber = 2, uiState, viewModel, onUseTestCardClick)
+            Step_VaultCard(uiState, viewModel, onUseTestCardClick)
         }
-
-        val authChallenge = uiState.authChallenge
-        if (authChallenge != null) {
-            Step_PresentAuthChallenge(stepNumber = 3, authChallenge, uiState, viewModel)
-        } else if (uiState.isVaultCardSuccessful) {
-            Step_CreatePaymentToken(stepNumber = 3, uiState, viewModel)
+        if (uiState.isVaultCardSuccessful) {
+            Step_CreatePaymentToken(uiState, viewModel)
         }
-
-        if (uiState.isVaultWith3DSSuccessful) {
-            (uiState.refreshSetupTokenState as? ActionState.Success)?.value?.let { setupToken ->
-                InfoColumn(title = "UPDATED SETUP TOKEN") {
-                    CardSetupTokenView(setupToken = setupToken)
-                }
-            }
-            Step_CreatePaymentToken(stepNumber = 4, uiState, viewModel)
-        }
-
         Spacer(modifier = Modifier.size(contentPadding))
     }
 }
@@ -104,14 +86,13 @@ fun VaultSuccessView(cardVaultResult: CardVaultResult) {
 
 @Composable
 private fun Step_CreateSetupToken(
-    stepNumber: Int,
     uiState: VaultCardUiState,
     viewModel: VaultCardViewModel
 ) {
     Column(
         verticalArrangement = UIConstants.spacingMedium,
     ) {
-        StepHeader(stepNumber = stepNumber, title = "Create Setup Token")
+        StepHeader(stepNumber = 1, title = "Create Setup Token")
         EnumOptionList(
             title = stringResource(id = R.string.sca_title),
             stringArrayResId = R.array.sca_options,
@@ -135,7 +116,6 @@ private fun Step_CreateSetupToken(
 @ExperimentalMaterial3Api
 @Composable
 private fun Step_VaultCard(
-    stepNumber: Int,
     uiState: VaultCardUiState,
     viewModel: VaultCardViewModel,
     onUseTestCardClick: () -> Unit
@@ -144,7 +124,7 @@ private fun Step_VaultCard(
     Column(
         verticalArrangement = UIConstants.spacingMedium,
     ) {
-        StepHeader(stepNumber = stepNumber, title = "Vault Card")
+        StepHeader(stepNumber = 2, title = "Vault Card")
         CardForm(
             cardNumber = uiState.cardNumber,
             expirationDate = uiState.cardExpirationDate,
@@ -172,44 +152,14 @@ private fun Step_VaultCard(
 
 @ExperimentalMaterial3Api
 @Composable
-private fun Step_PresentAuthChallenge(
-    stepNumber: Int,
-    authChallenge: CardAuthChallenge,
-    uiState: VaultCardUiState,
-    viewModel: VaultCardViewModel
-) {
-    val context = LocalContext.current
-    Column(
-        verticalArrangement = UIConstants.spacingMedium,
-    ) {
-        StepHeader(stepNumber = stepNumber, title = "Complete Auth Challenge")
-        ActionButtonColumn(
-            defaultTitle = "PRESENT AUTH CHALLENGE",
-            successTitle = "AUTH CHALLENGE COMPLETE",
-            state = uiState.authChallengeState,
-            onClick = {
-                context.getActivityOrNull()?.let { viewModel.presentAuthChallenge(it, authChallenge) }
-            }
-        ) { state ->
-            when (state) {
-                is CompletedActionState.Failure -> ErrorView(error = state.value)
-                is CompletedActionState.Success -> CardVaultResultView(result = state.value)
-            }
-        }
-    }
-}
-
-@ExperimentalMaterial3Api
-@Composable
 private fun Step_CreatePaymentToken(
-    stepNumber: Int,
     uiState: VaultCardUiState,
     viewModel: VaultCardViewModel
 ) {
     Column(
         verticalArrangement = UIConstants.spacingMedium,
     ) {
-        StepHeader(stepNumber = stepNumber, title = "Create Payment Token")
+        StepHeader(stepNumber = 3, title = "Create Payment Token")
         ActionButtonColumn(
             defaultTitle = "CREATE PAYMENT TOKEN",
             successTitle = "PAYMENT TOKEN CREATED",